app/requirements.txt
python-dotenv==0.12.0
requests==2.23.0
ovh==0.5.0
